Split DefaultNavbarItem into Mobile/Desktop subcomponents
