--- packages/docusaurus-theme-classic/src/theme/NavbarItem/DefaultNavbarItem.tsx
@@ -1,66 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import React, {type ReactNode} from 'react';
-import clsx from 'clsx';
-import NavbarNavLink from '@theme/NavbarItem/NavbarNavLink';
-import type {
-  DesktopOrMobileNavBarItemProps,
-  Props,
-} from '@theme/NavbarItem/DefaultNavbarItem';
-
-function DefaultNavbarItemDesktop({
-  className,
-  isDropdownItem = false,
-  ...props
-}: Props) {
-  const element = (
-    <NavbarNavLink
-      className={clsx(
-        isDropdownItem ? 'dropdown__link' : 'navbar__item navbar__link',
-        className,
-      )}
-      isDropdownLink={isDropdownItem}
-      {...props}
-    />
-  );
-
-  if (isDropdownItem) {
-    return <li>{element}</li>;
-  }
-
-  return element;
-}
-
-function DefaultNavbarItemMobile({
-  className,
-  isDropdownItem,
-  ...props
-}: DesktopOrMobileNavBarItemProps) {
-  return (
-    <li className="menu__list-item">
-      <NavbarNavLink className={clsx('menu__link', className)} {...props} />
-    </li>
-  );
-}
-
-export default function DefaultNavbarItem({
-  mobile = false,
-  position, // Need to destructure position from props so that it doesn't get passed on.
-  ...props
-}: Props): ReactNode {
-  const Comp = mobile ? DefaultNavbarItemMobile : DefaultNavbarItemDesktop;
-  return (
-    <Comp
-      {...props}
-      activeClassName={
-        props.activeClassName ??
-        (mobile ? 'menu__link--active' : 'navbar__link--active')
-      }
-    />
-  );
-}
