+++ packages/docusaurus-theme-classic/src/theme/NavbarItem/DefaultNavbarItem/Desktop/index.tsx
@@ -1,0 +1,34 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import React, {type ReactNode} from 'react';
+import clsx from 'clsx';
+import NavbarNavLink from '@theme/NavbarItem/NavbarNavLink';
+import type {Props} from '@theme/NavbarItem/DefaultNavbarItem/Desktop';
+
+export default function DefaultNavbarItemDesktop({
+  className,
+  isDropdownItem = false,
+  ...props
+}: Props): ReactNode {
+  const element = (
+    <NavbarNavLink
+      className={clsx(
+        isDropdownItem ? 'dropdown__link' : 'navbar__item navbar__link',
+        className,
+      )}
+      isDropdownLink={isDropdownItem}
+      {...props}
+    />
+  );
+
+  if (isDropdownItem) {
+    return <li>{element}</li>;
+  }
+
+  return element;
+}
